pyproject.toml
[tool.poetry]
name = "requiem-blog"
version = "0.1.0"
description = ""
authors = ["pelaezluis <luispelaezbedoya@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.92.0"
alembic = "^1.9.4"
sqlmodel = "^0.0.8"
python-jose = "^3.3.0"
cryptography = "^39.0.1"
SQLAlchemy = "1.4.40"
fastapi-async-sqlalchemy = "^0.3.12"
uvicorn = "^0.20.0"
datetime = "^5.1"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
bcrypt = "^4.0.1"
pymysql = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
